pyproject.toml
[tool.poetry]
name = "chat-buddy"
version = "0.1.0"
description = ""
authors = ["Eduard Vlasov <driver.edro@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.13.3"
python-telegram-bot = "^21.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
